Use non-blocking sends
If send operation blocked, this is most likely because the receiver is
not processing messages.

If we block because the send blocked, then we cause senders to us
to eventually block also.

It is better to drop packets when the queue is full, so other tasks can
continue to run.